requirements.txt
discord.py>=1.7.0
pytube>=11.0.1
pynacl>=1.4.0
detectlanguage>=1.5.0
deep_translator>=1.5.0
langdetect>=1.0.9
alt-profanity-check>=0.24.0
requests>=2.25.1
python-dotenv>=0.17.0
dropbox>=11.21.0
pycryptodome>=3.11.0
google-search-results>=2.4.0
pyyaml>=6.0
random-word>=1.0.7
python-Levenshtein>=0.12.2
fuzzywuzzy>=0.18.0
dice>=3.1.2